test(rtc-topology): mirror service behavior by owner
